requirements.txt
pydantic
authlib
python-dotenv
tomli
aiohttp
mcp

build

pytest
pytest-asyncio
pytest-aiohttp
pytest-cov
coverage
fastapi
uvicorn
pyyaml